requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.2
django-admin==2.0.2
django-crispy-forms==2.0
django-excel-response2==3.0.5
django-six==1.0.5
django-social-auth==0.7.28
excel-base==1.0.4
httplib2==0.22.0
idna==3.4
isoweek==1.3.3
oauth2==1.9.0.post1
oauthlib==3.2.2
psycopg2==2.9.6
pycparser==2.21
PyGithub==1.58.1
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-openid==2.2.5
python3-openid==3.2.0
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
screen==1.0.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
TimeConvert==3.0.13
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
wrapt==1.15.0
xlwt==1.3.0
